Fix pen pointer capture in iframe
When set/release/check pointercapture, we checked whether it's touch
type pointer. This is because touch like events will route to local main
frame's PME, while mouse like events will route event to hit tested frame.
However, it didn't check whether pen type events are send like touch.
This CL makes set/release/has pointercapture always check whether pointer
is active in the capture element's frame, if not, then check the local main
frame. So for touch like events, pointer active state is not available in
inner frame and we will always fall into check local main frame.
For mouselike events, it will either is activating in the inner frame.
If not, when checking main frame, IsPointerIdActiveOnFrame will return false
because the main frame's element_under_pointer_ is not match.

The test has an expected file for the second test failing on 'pen'. It's
PEM.GetMouseCaptureTarget returns the target for kMouseId, but for layout
test, pen events is send as mouse-like but has different id.
It's unnecessary to fix this now as we have direct manipulation enabled
on all platform, we're not going to have mouse-like pen event for capturing.
It'll be fixed once we have pen event on test driver work as current pen
flow.

Bug: 970595
Change-Id: I7f9f390d1526eefee4657f7b859edf5ad4d163ee