Optimize cubic bezier solver.
Newton's method requires an accurate guess to get good convergence.
This patch introduces a linear interpolation of a crude spline as a
starting point for the iterative solver. Previously the starting point
t = x was used.  The patch also introduces a tighter tolerance for
Newton's method than the fallback, and is justified by Newton's method
having quadratic convergence characteristics. Speed testing indicates
that the solver is roughly 50% faster, since the number of iterations
is typically less even at the higher accuracy requirement.

Bug: 827560
Change-Id: I89e3cc9061823269aabb16697e2d167751aa52b1