[LayoutNG] Fix DCHECK(NeedsLayout) for inline oof fixed container.
NGOutOfFlowLayoutDescendant.inline_container was not being set
correctly for an Element if:
- There is an OOF containing block between Element and
its Container(). This can only happens if Element is
position:fixed.

This caused Element not to be laid out after it got dirty (but
its containing block did not).

The fix is to set inline_container when Element's
NGOutOfFlowDesecendantCandidate gets propagated up the OOF
containing block chain.

For a while, I was afraid that this fundamentally broke
the concept of inline_container. I could not come up
with a counterexample that broke this fix.

Bug: 974760
Change-Id: Id16a0057f0aefe183c30c53244dd5c46108f093c